// src/interfaces/select-option.ts
/**
 * @Author: sarahnouh
 * @Date:   2019-04-10T06:02:57+02:00
 * @Last modified by:   sarahnouh
 * @Last modified time: 2019-04-10T06:05:12+02:00
 */
export interface optionType {
  /**
   *the value of the select option
   *@type string
   */
  value: string;
  /**
   *the label of the select option
   *@type string
   */
  label: string;
}