// src/interfaces/range.ts
/**
 * @Author: sarahnouh
 * @Date:   2019-04-10T06:00:26+02:00
 * @Last modified by:   sarahnouh
 * @Last modified time: 2019-04-13T19:07:45+02:00
 */
export interface Range {
  /**
   *The minimum value for the range
   *@type number
   */
  max: number;
  /**
   *The maximum value for the range
   *@type number
   */
  min: number;
}